+++ src/utils/getOutputPath.ts
@@ -7,6 +7,6 @@ import path from 'path';
 import { Config } from '../types';
 
 export const getOutputFilePath = (config: Config, file: string) => {
-  const { dir, name } = path.parse(config.outputFilePath || '');
+  const { dir, name } = path.parse(config.output || '');
   return path.join(dir, name, file);
 };